[networks] Add copyright headers
